DS-36 remove unused fields, rename abstract class
